Move fetch function to its own module
